+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.9.5 - 2017-10-31
+ * Allow strings to be passed in to `MessageType::setBccRecipients` and `MessageType::setCcRecipients` as well as their
+  functions to add recipients, just like we do for To Recipients
+
 ## 0.9.4 - 2017-08-29
  * Fixed some bugs with the usage of ensureIsArray